backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart>=0.0.7
pydantic-settings==2.6.1
